Test de Inscripcion Terminado

+++ msvc-inscripcion/src/main/java/com/edutech/msvc/inscripcion/services/InscripcionServiceImpl.java
@@ -74,6 +74,9 @@ public class InscripcionServiceImpl implements InscripcionService{
 
     @Override
     public void delete(Long id){
+        if (!inscripcionRepository.existsById(id)){
+            throw new InscripcionException("Inscripcion con id " + id + " no existe");
+        }
         this.inscripcionRepository.deleteById(id);
     }
